Add dry-run option to glossary

+++ packages/parser/src/commands/glossary.js
@@ -12,6 +12,7 @@ const {
 } = require("../lib.js");
 
 async function glossary(options) {
+  options.dryRun && console.log("\n* Dry run enabled *\n");
   let glossaryContent = "";
   const termsFiles = await getFiles(options.termsDir, options.noParseFiles)
   const termsData = await preloadTerms(termsFiles);
@@ -25,8 +26,15 @@ async function glossary(options) {
     const glossaryTerm = getGlossaryTerm(term, relativePath);
     glossaryContent = glossaryContent + glossaryTerm;
   }
-  const glossaryFile = getOrCreateGlossaryFile(options.glossaryFilepath);
-  fs.writeFileSync(options.glossaryFilepath, glossaryFile+glossaryContent, "utf8");
+  if(options.dryRun) {
+    const outputFile = options.logOutputFile;
+    fs.writeFileSync(outputFile,
+      `\n! These changes will not be applied in the glossary file.` +
+      `\nShowing the output below:\n\n${(glossaryContent)}\n\n`, "utf8");
+  } else {
+    const glossaryFile = getOrCreateGlossaryFile(options.glossaryFilepath);
+    fs.writeFileSync(options.glossaryFilepath, glossaryFile+glossaryContent, "utf8");
+  }
   console.log(`\u2713 ${cleanTerms.length} terms found.`)
 };
